Integrate doom-style runtime keybindings into seat.zig
Runtime binding flow:
1. main.zig loads defaults.xkb_bindings
2. main.zig loads basket.zon (unbind/bind/disable_defaults)
3. Merges: defaults - unbinds + basket binds
4. Sets kwm.runtime_bindings before seat creation
5. seat.zig uses runtime_bindings iterator (fallback to config.zig)

New file:
- src/kwm/runtime_bindings.zig: Global storage + iterator for runtime bindings

Changes:
- seat.zig: Uses runtime_bindings iterator instead of config.xkb_bindings
- main.zig: loadRuntimeBindings() merges defaults + basket.zon
- Fixed F-key parsing (u32 not u8)
- Mode enum conversion between defaults.Mode and config.Mode

The doom philosophy is now fully wired:
- Basket ships with sane defaults (no config needed)
- delta.zon for chord trees (newcomer-friendly popup)
- basket.zon for combo overrides (ricer territory)

+++ src/config/basket_config.zig
@@ -253,7 +253,7 @@ fn parseKeysym(key: []const u8) ?u32 {
 
     // F keys
     if (key.len >= 2 and key[0] == 'F') {
-        const num = std.fmt.parseInt(u8, key[1..], 10) catch return null;
+        const num = std.fmt.parseInt(u32, key[1..], 10) catch return null;
         if (num >= 1 and num <= 12) {
             return Keysym.F1 + (num - 1);
         }
